requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
DateTime==4.3
# Django==3.2.4
django-admin==2.0.1
django-cors-headers==3.7.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-hosts==4.0
django-six==1.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djongo==1.3.6
dnspython==2.1.0
docutils==0.17.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
screen==1.0.1
shortuuid==1.0.1
six==1.16.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.5
xlwt==1.3.0
zope.interface==5.4.0
